Huge changes

+++ industrial/swcpu/common/headers/arithm.h
@@ -1,4 +1,4 @@
 #include <stdint.h>
 
-int32_t str_to_num(const char str[]);
-void    num_to_str(int32_t num, char str[]);
+double  str_to_num(const char str[]);
+void    num_to_str(double num, char str[]);
